+++ fox-devices/_classes/device-model.class.js
@@ -15,14 +15,11 @@ class StoredDevice
 		this.model = validationTasks.readReferenceStringProperty(inputDeviceObj, "model", deviceSettings.listRioModelTypes);
 		this.name = validationTasks.readStringProperty(inputDeviceObj, "name", null, this.constructor.name);
 		this.desc = validationTasks.readStringProperty(inputDeviceObj, "desc", "", this.constructor.name);
-		
-		this.ipAddress = validationTasks.readStringProperty(inputDeviceObj, "ipAddress", null, this.constructor.name);
-		// Validate IP format.
-		
+		this.ipAddress = validationTasks.readIpAddressProperty(inputDeviceObj, "ipAddress");
 		this.username = validationTasks.readStringProperty(inputDeviceObj, "username", "", this.constructor.name);
 		this.password = validationTasks.readStringProperty(inputDeviceObj, "password", "", this.constructor.name);
 		this.isEnabled = validationTasks.readBooleanProperty(inputDeviceObj, "isEnabled", null, this.constructor.name);
-		this.macAddress = validationTasks.readStringProperty(inputDeviceObj, "macAddress", "", this.constructor.name);
+		this.macAddress = validationTasks.readStringProperty(inputDeviceObj, "macAddress", null, this.constructor.name);
 		this.isDeleted = validationTasks.readBooleanProperty(inputDeviceObj, "isDeleted", false, this.constructor.name);
 		this['__modified'] = validationTasks.readNumberProperty(inputDeviceObj, "__modified", 0, this.constructor.name);